feat: convert to hyphenenated prop on native
