+++ morph/react-dom/get-style-for-property.js
@@ -1,4 +1,9 @@
-import { getProp, getScopedCondition, isSlot } from '../utils.js'
+import {
+  getProp,
+  getScopedCondition,
+  isSlot,
+  maybeMakeHyphenated,
+} from '../utils.js'
 import { maybeAddFallbackFont } from '../fonts.js'
 
 export default (node, parent, code) => {
@@ -86,17 +91,6 @@ export default (node, parent, code) => {
 
 const maybeAsVar = (prop, code) =>
   code ? asVar(prop) : maybeMakeHyphenated(prop.value)
-
-const maybeMakeHyphenated = value => {
-  if (/^[a-zA-Z]+$/.test(value)) {
-    let splitStrings = value.split(/(?=[A-Z])/)
-    if (splitStrings.length > 1) {
-      return splitStrings.map(string => string.toLowerCase()).join('-')
-    }
-  }
-
-  return value
-}
 
 const asVar = prop => `'var(--${prop.name})'`
