+++ morph/utils.js
@@ -545,3 +545,14 @@ export const hasRowStyles = node =>
   node.properties.some(
     prop => prop.name.match(/^row/) && prop.name !== 'rowHeight'
   )
+
+export const maybeMakeHyphenated = value => {
+  if (/^[a-zA-Z]+$/.test(value)) {
+    let splitStrings = value.split(/(?=[A-Z])/)
+    if (splitStrings.length > 1) {
+      return splitStrings.map(string => string.toLowerCase()).join('-')
+    }
+  }
+
+  return value
+}
